+++ CHANGELOG.md
@@ -27,7 +27,7 @@
 - Added support for `Resampler.asfreq`.
 - Added support for `resample` frequency `W`, `ME`, `YE` with `closed = "left"`.
 - Added support for `DataFrame.rolling.corr` and `Series.rolling.corr` for `pairwise = False` and int `window`.
-- Added support for string time-based `window` for `Rolling`.
+- Added support for string time-based `window` and `min_periods = None` for `Rolling`.
 
 #### Improvements